Move lobby component into game page folder
