+++ client/src/app/app-routing.module.ts
@@ -3,7 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 import { HomepageComponent } from './homepage/homepage.component';
 import { NewGameCodeScreenComponent } from './new-game-code-screen/new-game-code-screen.component';
 import {EnterGameCodeScreenComponent} from './enter-game-code-screen/enter-game-code-screen.component';
-import {LobbyComponent} from './lobby/lobby.component'
+import {LobbyComponent} from './game-page/lobby/lobby.component';
 import { ChatComponent } from './game-page/chat/chat.component';
 import { VotingPageComponent } from './game-page/voting-page/voting-page.component';
