+++ client/src/app/game-page/lobby/lobby.component.ts
@@ -1,6 +1,6 @@
 import { Component} from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
-import { GameServerService } from '../game-server.service';
+import { GameServerService } from '../../game-server.service';
 import { Router } from '@angular/router';
